Added annotation validation

+++ src/main/java/com/epam/project/service/impl/UserServiceImpl.java
@@ -1,6 +1,8 @@
 package com.epam.project.service.impl;
 
+
 import com.epam.project.mapper.UserMapper;
+import com.epam.project.model.dto.UserCreateDto;
 import com.epam.project.model.dto.UserDto;
 import com.epam.project.model.entity.User;
 import com.epam.project.repository.UserRepository;
@@ -11,12 +13,15 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 @Service
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final UserMapper userMapper;
+    private static final Logger logger = Logger.getLogger(UserServiceImpl.class.getName());
 
     @Override
     public List<UserDto> findAll() {
@@ -24,23 +29,31 @@ public class UserServiceImpl implements UserService {
         return userMapper.toDtos(users);
     }
 
-    @Override
-    public UserDto findById(UUID id) {
-        User user = userRepository.findById(id)
-                .orElseThrow(() -> new UsernameNotFoundException("There is no user with ID= " + id + " in database"));
-        return userMapper.toDto(user);
-    }
 
     @Override
-    public UserDto save(UserDto userDto) {
-        User user = userMapper.toEntity(userDto);
-        user = userRepository.save(user);
-        return userMapper.toDto(user);
+    public UserDto findUserById(UUID id) {
+        User user = userRepository.findUserById(id);
+        if (user != null) {
+            return userMapper.toDto(user);
+        }
+        return null;
     }
+
+
+        @Override
+    public UserCreateDto save(UserCreateDto userCreateDto){
+        User user = userMapper.toCreateEntity(userCreateDto);
+        user = userRepository.save(user);
+        UserCreateDto newUser = userMapper.toCreateDto(user);
+        logger.log(Level.INFO, "User saved to database");
+        return newUser;
+    }
+
+
+
 
     @Override
     public UserDto update(UserDto userDto) {
-//        if (!userRepository.existsById(userDto.getId())) {
         if (!userRepository.existsById(userDto.getId())) {
             throw new UsernameNotFoundException("There is no user with ID= " + userDto.getId() + " in database");
         }
@@ -51,6 +64,14 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteById(UUID id) {
-        userRepository.deleteById(id);
+        User user = userRepository.findUserById(id);
+        if (user != null) {
+            userRepository.deleteById(id);
+            logger.log(Level.INFO, "User with id: " + id + " deleted");
+        } else {
+            throw new UsernameNotFoundException("User with id: " + id + " not available in database");
+        }
     }
+
+
 }
